chore: move services and pipes to folders
