+++ projects/ch-web/src/app/dash-board/dash-board.component.ts
@@ -4,10 +4,10 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Observable } from 'rxjs';
 import { AddWatcherComponent } from '../add-watcher/add-watcher.component';
-import { IdbService } from '../idb.service';
+import { IdbService } from '../services/idb.service';
 import { CenterForDay } from '../models/center-for-day';
 import { WatchInfo } from '../models/watch-info';
-import { WatchService } from '../watch.service';
+import { WatchService } from '../services/watch.service';
 
 
 @Component({
