+++ projects/ch-web/src/app/interceptors/http-progress.interceptor.spec.ts
