+++ projects/ch-web/src/app/interceptors/request-limiter.interceptor.ts
@@ -6,11 +6,11 @@ import {
   HttpInterceptor
 } from '@angular/common/http';
 import { EMPTY, from, iif, Observable, of } from 'rxjs';
-import { IdbService } from './idb.service';
-import { RequestQueue } from './models/request-queue';
+import { IdbService } from '../services/idb.service';
+import { RequestQueue } from '../models/request-queue';
 import { concatMap, flatMap, map, mergeMap, switchMap } from 'rxjs/operators';
 import { MatDialog } from '@angular/material/dialog';
-import { CustomAlertComponent } from './shared/custom-alert/custom-alert.component';
+import { CustomAlertComponent } from '../shared/custom-alert/custom-alert.component';
 
 @Injectable()
 export class RequestLimiterInterceptor implements HttpInterceptor {
