+++ projects/ch-web/src/app/search/use-district/use-district.component.ts
@@ -2,7 +2,7 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Observable, of, Subject } from 'rxjs';
 import { debounceTime, map, startWith, takeUntil, tap } from 'rxjs/operators';
-import { CowinDataService } from '../../cowin-data.service';
+import { CowinDataService } from '../../services/cowin-data.service';
 import { CenterForDay } from '../../models/center-for-day';
 import { CenterForWeek } from '../../models/center-for-week';
 import { District } from '../../models/district';
