+++ projects/ch-web/src/app/services/progress-bar.service.spec.ts
