+++ projects/ch-web/src/app/settings/settings.component.html
@@ -8,7 +8,7 @@
       <div class="mx-5 my-5 p-5 bg-green-100 rounded-xl">
           <article>
             Periodic Background Sync enables web applications to periodically synchronize data in the background.
-            <a href="https://web.dev/periodic-background-sync">More details.</a>
+            <a class="text-blue-700 hover:underline" href="https://web.dev/periodic-background-sync">More details.</a>
           </article>
       </div>
 </div>
